more debugging

+++ mashpia.com/public/rank_ceremony/index.php
@@ -27,7 +27,7 @@ function createZip($files, $images, $filename) {
     }
     foreach ($images as $img) {
         $zip->addFile($img);
-        unlink($img);
+        if (is_file($img)) unlink($img);
     }
     $zip->close();
 }
